requirements-test.txt
# Python testing dependencies for PolicyEngine

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0  # Parallel test execution

# Coverage reporting
pytest-cov>=4.1.0
allure-pytest>=2.13.0

# HTTP client for API testing
requests>=2.31.0
httpx>=0.24.0  # Async HTTP client

# gRPC client (if testing gRPC endpoints)
grpcio>=1.56.0
grpcio-tools>=1.56.0

# YAML parsing (for PolicyDomain files)
PyYAML>=6.0

# Utilities
python-dotenv>=1.0.0  # Environment variable management
